+++ packages/browser-wallet-api/src/wallet-api.ts
@@ -19,6 +19,7 @@ import {
     EventType,
     SchemaWithContext,
     SchemaType,
+    SmartContractParameters,
 } from '@concordium/browser-wallet-api-helpers';
 import EventEmitter from 'events';
 import type { JsonRpcRequest } from '@concordium/common-sdk/lib/providers/provider';
@@ -114,7 +115,7 @@ class WalletApi extends EventEmitter implements IWalletApi {
         accountAddress: string,
         type: AccountTransactionType,
         payload: AccountTransactionPayload,
-        parameters?: Record<string, unknown>,
+        parameters?: SmartContractParameters,
         schema?: string | SchemaWithContext,
         schemaVersion?: SchemaVersion
     ): Promise<string> {