+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   Added section for `requestIdProof` to README.
 
+### Fixed
+
+-   Type for smart contract parameters in sendTransaction.
+
 ## 2.2.0
 
 ### Added